// src/app/tours/shared/models/tour.ts
import { Currency } from '../../../shared/models/currency';
import { TourSurcharge } from './toursurcharge';
import { DepartureSchedule } from './departureschedule';
import { Departure } from './departure';
import { TourItineraryInfo } from './touritineraryinfo';
import { Vehicle } from './vehicle';
export interface ITour {
  TourDepartureScheduleId: number;
  Price: number;
  DiscountPrice: number;
  Currency: Currency;
  ChildrenPrice: number;
  InfantPrice: number;
  DepartureDate: Date;
  SiteId: number;
  TourSurcharge: TourSurcharge;
  DepartureSchedule: DepartureSchedule[];
  TourId: number;
  TourTypeId: number;
  TourCode: string;
  Name: string;
  Description: string;
  PhotoUrls: string[];
  LogoUrl: string;
  FreeVacancies: number;
  ViewCount: number;
  IsDomestic: boolean;
  RatingCount: number;
  CommentCount: number;
  OtherInformation: string;
  Included: string;
  Excluded: string;
  Departure: Departure;
  TourItineraryInfo: TourItineraryInfo[];
  PhoneNumber: string;
  IsLiked: boolean;
  NumberOfSeat: number;
  NumOfDays: number;
  NumOfNights: number;
  Vehicle: Vehicle[];
  Vehicles: Vehicle[];


}
export class Tour implements ITour {
  TourDepartureScheduleId: number;
  Price: number;
  DiscountPrice: number;
  Currency: Currency;
  ChildrenPrice: number;
  InfantPrice: number;
  DepartureDate: Date;
  SiteId: number;
  TourSurcharge: TourSurcharge;
  DepartureSchedule: DepartureSchedule[];
  TourId: number;
  TourTypeId: number;
  TourCode: string;
  Name: string;
  Description: string;
  PhotoUrls: string[];
  LogoUrl: string;
  FreeVacancies: number;
  ViewCount: number;
  IsDomestic: boolean;
  RatingCount: number;
  CommentCount: number;
  OtherInformation: string;
  Included: string;
  Excluded: string;
  Departure: Departure;
  TourItineraryInfo: TourItineraryInfo[];
  PhoneNumber: string;
  IsLiked: boolean;
  NumberOfSeat: number;
  NumOfDays: number;
  NumOfNights: number;
  Vehicle: Vehicle[];
  Vehicles: Vehicle[];
}